Add files via upload
This has some work done with isolation forests, which works well. With our current knowledge, it does not seem like it will outperform a transformer, but at the time it was a promising venture. This also showcases some basic conversion of the heading of the ship to velocity vectors, should we need it later.