Add files via upload
This file holds the general structure of a transformer, located at the bottom. Perhaps more important is the structure for defining paths. The paths were created by sorting by mmsi and timePassed, and then incrementing the path counter when a new ship passed information or a time of more than an hour passed. The paths of size <8 were eliminated, so only paths of a useable size were kept.