Add files via upload
This is a newly tested isolation forest, right now I have it as a backup if we cannot get our second part finished in time. I ran it on the preprocessed test data, and then used the transformer to create a path given the first few points of the test data. After the transformer created the path, I fed that directly into the prediction part of the isolation forest, which predicted all of the points were indeed non-anomalous.  This was unsurprising, but just to be sure, I made some modifications to the points, and after running again it predicted the changed points were anomalous. At the very least, this shows that the transformer is predicting the ship routes decently correctly (assuming that the route I picked was non-anomalous).